+++ jdaviz/configs/default/plugins/tools.py
@@ -1,7 +1,7 @@
 from glue_jupyter.bqplot.profile import BqplotProfileView
+
 from jdaviz.core.tools import SinglePixelRegion
 from jdaviz.core.marks import PluginLine
-
 
 __all__ = ['ProfileFromCube']
 
@@ -16,11 +16,8 @@ class ProfileFromCube(SinglePixelRegion):
         self._data = None
 
     def _reset_profile_viewer_bounds(self):
-        pv_state = self._profile_viewer.state
-        pv_state.x_min = self._previous_bounds[0]
-        pv_state.x_max = self._previous_bounds[1]
-        pv_state.y_min = self._previous_bounds[2]
-        pv_state.y_max = self._previous_bounds[3]
+        self._profile_viewer.set_limits(
+            y_min=self._previous_bounds[2], y_max=self._previous_bounds[3])
 
     def activate(self):
         self.viewer.add_event_callback(self.on_mouse_move, events=['mousemove', 'mouseleave'])
@@ -34,8 +31,7 @@ class ProfileFromCube(SinglePixelRegion):
             self._mark = PluginLine(self._profile_viewer, visible=False)
             self._profile_viewer.figure.marks = self._profile_viewer.figure.marks + [self._mark, ]
         # Store these so we can revert to previous user-set zoom after preview view
-        pv_state = self._profile_viewer.state
-        self._previous_bounds = [pv_state.x_min, pv_state.x_max, pv_state.y_min, pv_state.y_max]
+        self._previous_bounds = self._profile_viewer.get_limits()
         super().activate()
 
     def deactivate(self):